企业微信消息推送

+++ bush-message/message-api/src/main/java/com/bushro/message/properties/DingTalkCorpConfig.java
@@ -27,6 +27,6 @@ public class DingTalkCorpConfig extends Config {
     /**
      * 应用agentId
      */
-    private Integer agentId;
+    private Long agentId;
 
 }
